fix rabbit madness

+++ freg-nox/blocks.h
@@ -457,7 +457,7 @@ class Bush : public Active, public Inventory {
 class Rabbit : public Animal {
 	Q_OBJECT
 
-	float Attractive(int sub) const;
+	short Attractive(int sub) const;
 
 	public:
 	int  Kind() const;
